+++ docs/Changelog.md
@@ -5,6 +5,7 @@
 - vtm-desktop-lwjgl module [#714](https://github.com/mapsforge/vtm/pull/714)
 - vtm-desktop-lwjgl3 module [#717](https://github.com/mapsforge/vtm/pull/717)
 - Fix marker touch events [#723](https://github.com/mapsforge/vtm/issues/723)
+- Android 10 compatibility [#728](https://github.com/mapsforge/vtm/issues/728)
 - Many other minor improvements and bug fixes
 - [Solved issues](https://github.com/mapsforge/vtm/issues?q=is%3Aclosed+milestone%3A0.12.0)